Add year selector to updates page
